Update 2.2 -- September 29, 2018 9:48PM (UTC-06:00)
1. Login System Overhaul
  Login system now allows users to login with their email or username.
  Error handling is drastically improved.
  Header now displays error when incorrect login credentials are used. 
  Now using an include (login_handler.php) on forum action instead of 
embedding the code
  
2. Designs
  The login page looks much better and is more user friendly.
  
  
This update was mostly to improve the user(s) experience when logging in 
and creating a much more effecient system of code. 
I think my work has paid off and it looks extreamly professional and has 
improved the look over the previous update.
I still need to work on the remember me feature and adding in too many 
login attempt (reCAPTCHA) perhaps?